requirement.txt
google-generativeai
langchain
streamlit
streamlit-chat
tiktoken
python-dotenv